Switch to vite as a build system
Using react-scripts and create-react-app are
apparently old and have a lot of dependencies
with vulnerabilities. There was also an issue
with using them an having an up-to-date typescript.

Use vite instead which significiantly lowers
the amount of transitive deps we need.

This also fixes tests in the CI.
