+++ .github/workflows/ci.yml
@@ -15,4 +15,5 @@ jobs:
     - uses: actions/setup-node@v4
     - run: npm install
     - run: npm run lint
+    - run: npm run test
     - run: npm run build
