+++ index.html
@@ -8,11 +8,12 @@
       name="description"
       content="Visualizing Linux Kernel BPF verifier log to help BPF programmers with debugging verification failures"
     />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="manifest.json" />
     <title>BPF Verifier Visualizer</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
