+++ package.json
@@ -3,25 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/dom": "^10.4.0",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^13.5.0",
-    "@types/node": "^16.18.126",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
+    "react-dom": "^19.1.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "serve": "serve -s build",
+    "start": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
     "test": "jest",
-    "eject": "react-scripts eject",
     "format": "prettier src/* --write",
     "lint": "eslint src/*"
   },
@@ -31,6 +20,7 @@
       "react-app/jest"
     ]
   },
+  "type": "module",
   "jest": {
     "transform": {
       "^.+\\.(ts|tsx)$": "ts-jest"
@@ -50,11 +40,19 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
     "@types/jest": "^30.0.0",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@typescript-eslint/eslint-plugin": "^8.37.0",
+    "@typescript-eslint/parser": "^8.37.0",
+    "@vitejs/plugin-react": "^4.7.0",
     "jest": "^30.0.4",
     "jest-environment-jsdom": "^30.0.4",
-    "prettier": "3.6.2",
-    "serve": "^14.2.4",
-    "ts-jest": "^29.4.0"
+    "prettier": "^3.6.2",
+    "ts-jest": "^29.4.0",
+    "typescript": "^5.8.3",
+    "vite": "^7.0.5"
   }
 }
