+++ src/App.test.tsx
@@ -2,7 +2,7 @@
  * @jest-environment jsdom
  */
 import "@testing-library/jest-dom";
-import { render, screen, createEvent, fireEvent } from "@testing-library/react";
+import { render, createEvent, fireEvent } from "@testing-library/react";
 import App from "./App";
 
 describe("App", () => {
