+++ vite.config.js
@@ -1,0 +1,17 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+
+export default defineConfig(() => {
+  return {
+    build: {
+      outDir: 'build',
+    },
+    plugins: [react()],
+    base: '/bpfvv/',
+    server: {
+      fs: {
+        strict: false, // allow url content loading
+      }
+    },
+  };
+});
